Refine the setup script and implement basic SELECT queries
Basic MBQL queries that simply select a number of fields can now run. This also
required a patch to Toucan to make it not clash with Datomic. This has been
reported upstream.

Set the classpath through `lein` rather than at runtime.

Load the Datomic driver as part of booting Metabase.