Skip ref attributes
"foreign keys" is an optional feature for drivers. There are a few things that
need to be figured out there, e.g. how to deal with cardinality/many refs. Refs
also behave differently in that their pull result is a map with :db/id rather
than a scalar, so far we were marking them as integer, which caused metabase to
try to build up a histogram, which would fail.