First steps in proper query generation (WIP)
Go from the initial extremely naive implementation to something that has at
least some smarts about interpreting MBQL and converting it to Datalog.

So far this supports :fields, :breakout, and :aggregation [[:count]].

One challenge here is to stick to the SQL semantics where rows can have nils for
selected fields. If you select fields using datalog's `:where [[?eid :attr ?a]]`
then this will at the same time function as a filter, which we don't want. (for
breakouts we do want this, as it also acts like a GROUP BY in that case,
although that does mean you don't get a NULL-group back).

You can solve this with pull expressions in the :find clause, but that doesn't
lend itself too well to incremental generation (although we may still go back to
this). Instead we only retrieve entity ids, and resolve the attributes later
with `datomic.api/entity`. This does mean we need to pass extra
information ("field lookups") from the native query generator to the part that
executes it, and it means we'll have to do extra effort to support "plain"
Datalog queries in Metabase's raw query interface.