Streamline REPL setup, add aggregation test, simplify field names
Fields that have the same prefix as the table name they belong to now drop the
prefix, so "name" in the "artist" table, instead of "artist/name".

Add a test to assert the aggregate count functionality.

Change the elaborate lein injection command line with a bit of ed scripting +
tools.deps aliases. Much nicer.