Implement foreign key support
This commit represents a good deal of work to implement foreign key (:fk->)
clauses in varies places in MBQL queries, e.g. :field, :breakout, :order-by,
:aggregates.

These map fairly straightforwardly to datalog binding clauses, with some
caveats. Note that we still have the situation where a :where clause implicitly
filters, which is fine for :breakout but not for :fields, so for these we still
follow the approach of resolving afterwards, although the current
approach *will* limit the result to entities where the at least the ref is
present (but perhaps not the referenced attribute).

Because refs can be multiple (cardinality/many) we expand the result to its
cartesian product, so each ref in a set generates a new output line.

This commit also implements several more aggregations, including min, max, sum,
stddev.

A good deal of tests from Metabase's query-processor suite have been converted.
Note that Metabase's test-data assumes it can dictate primary keys (or that they
are assigned starting from 1). In Datomic this is not the case, the transactor
assigns entity ids, all we can do is supply tempids.

The data interface used to import test data has been adapted to work around
this, by rewriting foreign key references to temp ids, this does mean that a lot
of the tests that come with Metabase need to be rewritten before they apply to
us.