Add special case for a count of zero
Datomic's aggregates return no results (zero rows) when no entities match,
rather than a result of zero (a single row with a single element: 0). This is
pretty annoying, and means that there are times where Metabase expects to get a
single number back, and gets nothing.

We'll have to fix this properly to also work with things like sum, but for now
I'm just adding a check for queries that only ask for a single :count, as those
are quite common when clicking around the Metabase UI.