Significantly rework query generation regarding outer joins
Before we had two main approaches to query generation: :breakout based queries
would generate datalog that pulled out the relevant fields through datalog
:where clauses, whereas :fields based queries only pulled ou the entity id, and
resolved the attributes in a second step, outside of the query.

This led to some problems because not all information was available during query
time, so e.g. filtering on a foreign key field would not work as expected.

Another thing that didn't work properly were joins on cardinality/many
references, we expanded these into their cartesian product outside the query,
but entities with zero referenced attributes would not be included in the
result.

This change unifies these approaches, by doing much more through datalog, and
using `get-else` (or a glorious `or-join` hack in the case of cardinality/many
attributes), to bind (unify) attributes to logic variables without influencing
the rows/entities included in the result.