Correctly emulate get-else for cardinality/many
Datomic does not support get-else for reference attribtutes with
:db.cardinality/many. We work around this by using an or-join that will will
bind either to the actual references, or to Long/MIN_VALUE when there are no
references.

The form we used before always bound to Long/MIN_VALUE, so you would always get
an extra row returned with a nil value, even whent there were actual values.

This fixes that by only binding to Long/MIN_VALUE when the attribute is missing
on the entity.