Update the "empty result count" hack
When doing a (count ...) query and the result is 0/empty, datomic will return an
empty result (zero rows) rather than a row with count being [0]. We work around
this specific edge cases by pasting in a [0] result, but this only works when
not doing grouping. If you instead query for "count of rows grouped by X" and
the table is empty, then the result should also be empty.